[FIX] hr_appraisal: send appraisal request by email
Issue:
======
Clicking on send by email on appraisal request will raise an error.

Steps to reproduce:
===================
- Create an appraisal request
- Click on send by email button

Origin of the issue:
====================
In the form view of hr_appraisal , we are displaying recepient_ids in a
`many2many_tags` widget which requires `display_name` field, according
to the code of `onchange` here
https://github.com/odoo/odoo/blob/671640cf595dcff6a827b676a04fbe12575a4909/addons/web/models/models.py#L949-L956,
We need the values of recepiend_ids to be a command.

Solution:
=========
Use `[Command.set(ids)]` instead of the list of ids directly to make it
compatible with onchange.

opw-3604104

closes odoo/enterprise#51507

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>