[FIX] account_invoice_extract: no auto extract on demo posted invoices
When checking if an invoice needs to be automatically sent to the OCR,
we weren't checking the state of the invoice. Only draft invoices should
be sent.

In practice, it wasn't much of an issue as new invoices are created in
the `draft` state, but for demo invoices, some of them are already in
the `posted` state. Those shouldn't be sent for digitization.

closes odoo/enterprise#51529

Signed-off-by: Florian Daloze (fda) <fda@odoo.com>