[IMP] sale_planning: prevent auto-planning of shifts linked to canceled SOs
Before this commit, planning will auto-plan shifts linked to canceled SOs.
This doesn't make sense as the service no longer needs to be provided.
In addition, the user cannot easily delete the related unplanned shift so
it will keep getting scheduled automatically every time the feature is used.

In this commit, when an SO is canceled, the related unplanned shift will be
automatically deleted (excluding shifts that are already scheduled). If the SO
is reset to quotation and confirmed again, a new shift will be generated to plan
 for the remaining hours.

task-3603886

closes odoo/enterprise#51742

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>