[FIX] timesheet_grid: order records via section and row fields
Steps:
- Install Timesheet app.
- Open timesheet.

Issue:
- Timesheet are not order the way it was before
grid owl conversion.

Cause:
- After grid view owl converion grid is fetching data
from server side and it was not taking any orderby into
account so it was getting data according to default order
on timesheets which is `date desc, id desc`.

Fix:
- Added a method to get order according to section and row
fields so it will display data according to user pref and
it'll be more convenient for users.

task-3610476

closes odoo/enterprise#53999

X-original-commit: 8d3fc95577536e01c863057b58795848dad922c7
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>