[IMP] knowledge: add FormStatusIndicator buttons
This commit aims at adding the status indicator buttons that are present in
most other Odoo Form views to Knowledge.
These indicators enables the user to know when changes has been taken into
account and can be saved inside the DB, either using the buttons or using other
methods.

This also enables the user to discard any changes he'd done to the article
quickly.
We also added an icon when all changes have been committed, which is added via
the extension of the FormStatusIndicator Component.

task-3609594

closes odoo/enterprise#54547

X-original-commit: b49c8fe6279d679a8fba4d8fe4f8008ab55c4bfc
Related: odoo/odoo#149797
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>