[FIX] website_sale_stock_renting: fix renting dates in tour
Before this commit, default dates were used for renting. Launching this
test just before midnight, start date and end dates are different,
although the remaining of the test relies on the fact that hours are
defined for the same day. This commit makes sure the renting start date
is the same as the end date.

Fixes build error 28629

closes odoo/enterprise#51633

X-original-commit: c49d8f17a8fe24621423c7fcc9c03777dcadac9d
Signed-off-by: Vallaeys Valentin (vava) <vava@odoo.com>