[FIX] account_followup: fallback reply-to if not set in template
A recent bugfix PR (https://github.com/odoo/enterprise/pull/50621) added
a new function `_get_email_reply_to` in aims to correctly fetch an
approriate reply-to email address if it was setup in the email template
used for the followup messages.

But inadvertently it introduced and unexpected behavior:
if the email template has no reply-to set, the fonction returns `False`.
In that scenario, when the final email gets prepared for sending out,
calling `IrMailServer.build_email` and reply-to is `False`, it falls back
to the the `email_from` encoded in the email.

So in pratice, instead of using the default catchall email address,
it will use the same email for the `From` and `Reply-To` email headers.

This is an issue, since replying to the followup email might not correctly
reroute the incoming email to the correct model (res.partner in this case).

How to reproduce:
1) Make sure that the default template for the followup emails has no
reply-to set up (should be the case on a fresh database). Set up a system
parameter for `Key -> mail.catchall.domain` with example
`Value -> mycompany.example.com`
2) Go to Accounting > Customers > Follow-up Reports
3) Select a record
4) Click the "Follow Up" button which will open up the wizard. Let's assume
that contact uses `"partner@email.com"` as email address and that current
internal user uses `"user@internal.com"`
5) Send out email with default values

Email being sent will show:

Before this fix:
Reply-To: `"User" <user@internal.com>`
To: `"Partner" <partner@email.com>`

After this fix:
Reply-To: `"Company" <catchall@mycompany.example.com>`
To: "Partner" `<partner@email.com>`

opw-3605881

closes odoo/enterprise#51725

X-original-commit: 3efb2715139713677f53e83487d3aa4140221257
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Joel Rodrigues Vitória (jorv) <jorv@odoo.com>