[FIX] sale_subscription: allow to cancel and set to quotation uninvoiced orders
Before this commit, when an order was a renewal of an invoiced order, it was
not possible to cancel it even if this renewal had no invoice.

It was also not possible to put it back to quote because the conditions was based
on the invoice_count value.

Taskid: 3527957
X-original-commit: 5e52d2c9d47f3fce9eccf7c7ad77cdbb429f298e
Part-of: odoo/enterprise#51813