[REF] l10n_mx_edi: addenda
The user had to choose ir.ui.view in a Many2one field on the partner
which was not very user friendly and caused
1) some performance issues
2) impossibility to set correct correct security permissions

Instead of this approach we create a dedicated model for addenda and
a many2one field on the partner to access them.
We also switch some standard addendas that were demo data to regular data.
For now we have: Bosh, Autozone.

We also add the possibily to set the addenda per invoice, whose default
will be taken from the partner. This settings is now in the Other Info tab.

We also add list/form view for accessing the addenda in the Accounting
Setting tab.

Remove the l10n_mx_edi_addenda_doc that is unused.

closes odoo/enterprise#51691

Task-id: 2969591
Related: odoo/upgrade#5445
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>