[FIX] sale_subscription: 'smart' confirmation of renewal
Case:
1) order with a renewal quotation
2) the parent order is churned (invoice not paid)
3) the customer pay and confirm the renewal quote

Before this commit, the parent remains churned and the renewal produces a log
with an event_type: 'new'.

After this commit, the parent is reopened, the churn log is cleaned and the quote is
a normal renewal with renewed state on the parent. The log analysis don't produce a
churn log anymore.

closes odoo/enterprise#51578

Taskid: 3527957
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>