[FIX] appointment: handle resource availability on last day
ISSUE

On an appointment time selection screen, on the front-end,
slots may appear on the last day that are not really free,
or not matching resource scheduled work hours. Selecting
those will throw a not Found error as they will not be
possible / available slots and verification will fail.

TO REPRODUCE

a) Work hours not taken into account

1. Create appt_type A with one resource R
2. Set work hours to 40h/week on resource R
3. Set max scheduled days on A to 14 days
4. Have a slot on A from 10.00 to 23.00 on the
same weekday as now. (Let us say it's monday morning)
5. Go to front-end: the last available day (monday)
will have slots until 23.00, despite the working hours
restricting them to max 17.00.

b) Bookings not taken into account. (rarer)

If appointment tz is such that NOW is in next day compared
to UTC, if some bookings/leaves exist on that next day (in UTC),
they will not be taken into account when computing slots.

FIX

As the last day's slots are computed for the whole day,
the unavailabilities should also be computed for it. At the
moment, they are only checked until (now + max days) for
work hours and (now + max days) at 23.59.59 for bookings. We
take instead time.max on that day (as per tz of appointment type!)
as max range to check the unavailabilities of those slots, if any.

TEST

Existing test_manage_leaves is modified to keep a minimal diff.

Task-3612857

closes odoo/enterprise#54211

X-original-commit: 16f936339f55361a916484c1c473e69a032888ad
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Noé Antoine (nan) <nan@odoo.com>