[FIX] hr_appraisal_survey: hide appraisal option in survey form view
In odoo/enterprise#38268, we have added a survey type to allow to simplify the
survey creation by only displaying relevant options for the type chosen. We
have also added the appraisal type in the hr_appraisal_survey module, and
hidden it in the survey views because that type is only relevant for the
appraisal module and we don’t want to see it in the survey app.
From 16.4, that type is no longer hidden in the app survey.
This fix solves the problem.

Technical notes:
Following the change in odoo/enterprise#37387 that makes "primary" the override
of the survey form that hide the appraisal type, the appraisal type is no longer
hidden in the survey app. This is because in the override, we add a class used
in css to hide that option (no longer present in survey because of the primary).

To avoid modifying view in stable, we target the "appraisal" option without that
added class.

Task-3613370

closes odoo/enterprise#52045

X-original-commit: df1485e866f2a8110fed963f93b8470aa69f8740
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>