[IMP] knowledge: allow dropping articles in the shared section
Purpose:
--------
Previously, the shared section allowed dropping an article only if it was
moved under another shared article or if the article being moved was
already a shared root article.

With this commit, dropping an article into the shared section is permitted
as long as this article is shared with at least one partner (excluding
the user initiating the move and internal permissions, but including
inherited permissions).

In the case where a partner had access to the moved article through access
inheritance, a new member will be created for this article so that this
partner keeps his read/write access on this article.

Task-3603579

closes odoo/enterprise#51598

Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>