[FIX] account_batch_payment: removes 20 payments per page limit in print
Steps to reproduce:
- Install Accounting
- Activate "Batch Payments" in Accounting settings
- Go to "Accounting / Customers / Payments"
- Create more than 20 payments
- Select all the payments and create a batch payment
- Confirm the batch payment and print it
=> Each page of the generated PDF only contains 20 payments,
letting a lot of white space at the bottom of the page between
each page.
If the customer names are very long and are displayed on several lines,
it can happen that the 1st page contains the 19 first payments (displayed
on the whole page) and the 2nd page only contains the 2Oth payment.
Then the next 20 payments are displayed on the 3rd page, leaving the 2nd
page almost empty.

opw-3597753

closes odoo/enterprise#53364

X-original-commit: 98e34c14e70d2a82933a7c00e84642bf6e8d3cd5
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>