[FIX] whatsapp: set template attachment while syncing
Before This Commit

Currently, when a template with an image header is synced into the odoo database
the actual uploaded image file disappears and instead on the odoo database, a
file named missing.jpg is synced which contains no data in the same. The same
happens with a file with a document heading. It syncs a file name missing.pdf
Also when the template is submitted the attachment is not sent properly.

After This Commit

When the template is synced set the header attachment instead of the missing
document.

Solution

When the template is synced we get the document_url of the template demo
document then get the data, mime type, and extension from that document_url
and set it in whatsapp.template header attachment.

Also send raw field instead of datas field of header attachment when submitting
the template to meta.

Task-3576712 (WhatsApp: Sync template attachments at sync, avoid missing)

X-original-commit: odoo/enteprrise@f9032dbba3637c3f4ef3858fe23b5b6c506011be
Part-of: odoo/enterprise#59987