[IMP] sale_subscription: check deferred revenue settings before cron
Before this commit, it was possible to launch the subscription invoice cron when
the deferred journal or account was not properly set in the settings. As a result,
the invoice was created in draft, the payment request was performed but the invoice
was not posted. Then the system was rollback. An email indicating the issue was
sent but it was not clear enough. This commit improve the email content and
prevent to launch the cron if the deferred revenues are not properly configured.

closes odoo/enterprise#51632

Opw-ticket: 3565556
Signed-off-by: Leonardo Pavan Rocha <lpr@odoo.com>
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>