[IMP] account_reports: column order
Some users prefer to have the order of their period comparison in ascending or
descending order. By default, most reports will be on descending order except
for the Trial balance that is in ascending.

This PR will introduce, in the comparison dropdown, a new selection input that
will change the order of the comparison column.

closes odoo/enterprise#51643

Task: 3586472
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>