[IMP] planning: resource group_expand on role
When the planning is grouped by resource, and that the user search
for a resource that has a specific role, we would like to display
every resource that has this role in order to ease shift encoding.
Thus this commit modifies the group expand on resource when searching
on role in order to diplay even empty lines.

taskid:3186636

Part-of: odoo/enterprise#41943