[IMP] sale_subscription: reset end_date when a closed order is reopened
Before this commit, when a churned order was reopened by portal payment,
the end_date was left untouched. The expiration cron could close it the
same day than the reopening. This commit remove the end date.

We need to remove it and notify because we can't track all the possibilities:

* order reaches the end date, auto close, customer pay to reopen it:
   --> we want to remove the end_date
* 2 years order is closed by customer after a few months, it was a mistake.
We want to reopen it but the end date should be preserved.

It is not possible to distinguish easily these 2 cases and the others.
As we want to keep the flow simple, we decide to let the salesperson decide.

opw-ticket 3614318

closes odoo/enterprise#51732

closes odoo/enterprise#52460

closes odoo/enterprise#52599

X-original-commit: b34a6636176a82aec0d26ead728526e38fd40063
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>