[IMP] account_avatax: avatax test connection pop up
Before this commit, the pop-up when testing the avatax connection was very
unclear for user. Indeed, we displayed a dictionary of information without
really telling the user that the connection was successful or not.

Now, when testing the connection we clearly indicate if the authentication was a
success or not, and we display all the information as a list.

closes odoo/enterprise#52189

Task: 3614808
X-original-commit: c48802caa938af1b0b5ddc768ae00230bbcafa20
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>