[FIX] documents_spreadsheet: Fix tour data (again)
The fix proposed in 5d19c678 was not thorough as the pivot formula
present in the demo sheet will call `name_get` on given record ids that
are likely not existing without the demo data (see [1]). This revision
adds a dedicated spreasheet file that goes around that problem and
limits the name_get calls to actual data (in this case, rely on the
country_ids).

[1] https://runbot.odoo.com/runbot/build/54397846

closes odoo/enterprise#55015

X-original-commit: 1401c5c8eeff38b037c5371e6ae6d8dbc700c8f7
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>