[FIX] l10n_pe_edi: always put invoice total in PayableAmount
Currently our generic UBL mechanism puts the invoice amount either in
`PayableAmount` or `PrepaidAmount` depending on what has already been
paid.

For Peru however, the `PrepaidAmount` is reserved for advances payments
(anticipos) and should not be used for anything else. That's why we
change the behavior to always put the total invoice amount in the
`PayableAmount` tag.

Follow-up on task-3415758

closes odoo/enterprise#51738

Signed-off-by: Laurent Smet (las) <las@odoo.com>