[FIX] appointment, *: adapt "availabilities" button
*: appointment_crm, knowledge

In commit[1] "Share availabilities" button was moved to the control
panel.

This broke the visual aspect of the button's contents mainly because
there wasn't enough space for it.

Since the revamp of the calendar in
https://github.com/odoo/odoo/pull/138670 we introduce a header for the
calendar.

In this commit we move the button between the header and the sidebar
toggler on desktop and to the bottom of the screen on mobile.
We also fix its broken elements.

[1]: d7f1b4bea7c7586a0d26015ed353b97dd3f18721

task-3617561
part of task-3575827

Part-of: odoo/enterprise#51755