[FIX] documents_spreadsheet: Fix tour
The tour `spreadsheet_create_list_view` would sometimes fail as the
`beforeunload` process of the spreadsheet action occured at the last
step of the test and is slow enough to be interrupted when the tour
ends.

This revision addresses this issue by adding an explicit last step
to let the spreadsheet action clean itself properly.

fixes runbot errors 21468

closes odoo/enterprise#51994

X-original-commit: 3253582e44a59c7b480174506b153d29a31c656c
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>