[IMP] planning: move some buttons under a dropdown
This PR moves the 'copy previous week' and 'auto plan' buttons under a
dropdown right next to the 'publish' button. These changes are done for
the visibility of the planning app.

Since it is impossible to prevent the gantt view's ControlPanel from
putting all the available buttons in a dropdown on mobile view, the
buttons had to be duplicated in the views. One set of buttons will
appear only when display = xl and the other set of buttons will appear
only when display < xl. The tour was also edited to add extra steps,
since clicking on the 'copy previous week' button now requires you to
click on the dropdown first.

task-3613692

closes odoo/enterprise#51821

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>