[IMP] account_reports: introduce horizontal split feature
In some countries (like India), accountants are really used to having the BS displayed in two halves, next to each other (assets / liabilites + equity), and we receive a lot of feedbacks from them about changing the layout of the report to have something more intuitive for them.

This PR adds a generic mechanism to allow splitting the lines of a report between two "pages" displayed next to each other. This behavior can be enabled by specifying on the report lines whether they need to go left or right when splitting the report. If any of the account.report.lines of the report defines a value in that field, the report is going to be displayed split by default (lines with no value in that field will be put on the left side by default ; so we always display them all). An item in the extra options filter allows disabling that mode.

For reports using dynamic lines instead of expressions, horizontal split can also be used, by forcing the 'horizontal_split' key into the options through a _custom_options_initializer.

Task 3618056

Part-of: odoo/enterprise#51783