[FIX] planning: make date changes impact correctly the recurrency
Steps:
- create a shift recurrency every week and save
- go back to the shift, and set another date
- select "edit: all shifts" and save

Issue:
All shifts are set to the same date.

Cause:
We find all shift of the recurrence and apply them the same values.

Fix:
Delete all the shifts to update consequently, and recreate them.

Limitations:
Say a recurrence with:
`repeat_interval`: 1
`repeat_unit`: "day"
`repeat_type`: "x_times"
`repeat_number`: 7
first occurence on monday.
In other words, a shift every day of one week.
The user changes tuesday's shits's date to wednesday.

1) `recurrence_udate`: "all":
First shift will be on wednesday, next on thursday, etc. until next
tuesday. (=7 shifts)
This is because we update the first shift and recreate the following
ones.

2) `recurrence_udate`: "subsequent":
Everything is going to be updated as expected, but the last shift will
not be recreated. This is because, to know how many shifts have to be
created in a recurrence, we convert the `repeat_number` into a max date.
It is computed as: the end date of the first shift + (number of occurence
 - 1) * period between shifts.
Here: monday + (7-1) * 1 day = sunday.

opw-3571081

closes odoo/enterprise#57969

X-original-commit: c01b3b629aefe8995ac937e88d52715927f6483d
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Audric Onockx (auon) <auon@odoo.com>