[IMP] account_accountant: specific tax prediction
The aim of this commit is to provide a method allowing to use tax
prediction in a more specific way.
It allows to fix the amount, amount_type and use_type.

Context:
It will be used by the edi's import to retrieve the most accurate tax
that fits those criteria.

Before this commit:
We were only able to retrieve previous tax but without giving specific
needs.

After this commit:
It's possible to retrieve the most probable tax matching the criteria.

closes odoo/enterprise#52445

Task-id: 3626501
Ticket-id: 3531473
X-original-commit: c681416553bce282e68698988a5d3f4193c8c9ec
Related: odoo/odoo#145594
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>