[FIX] account_accountant: wait for reset in a bank reco tour step
Currently there is the following problem in the tour.
In some step we reset the the matching of a line.
In the following step there is no trigger or extra_trigger waiting for the reset to finish.
Due to this we may switch to the list view before the reset is finished (2 steps after starting the reset).
If this happens the tour fails since the line we reset needs to be unmatched in the list view (to be able to modify it).

closes odoo/enterprise#51922

X-original-commit: fdac623bb97811e35fcc077db40b4c35d237c12a
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Sven Führ (svfu) <svfu@odoo.com>