[FIX] knowledge: update ID references stored in the body of the copy
When the user copies an article, the system copies the body of the article
'as-is' without updating the ID references stored within it. As a result,
the embedded views of article items stored in the body of the copy will
list the article items of the original article and not its own article
items.

This commit will address that issue by updating the context of the
embedded views stored in the body of the copy. This will ensure that
the embedded views listing the article items of the original article
will now list the article items of the copy.

Steps to reproduce the issue:
1. Create an article
2. Insert an embedded view of article items (with /kanban or /list)
3. Insert a few articles
4. Create a copy of the article (using the "Create a Copy" button)

=> The embedded views of article items list the article items of the
original article.

TO BE: The embedded views of article items should list the article items
of the newly created article.

task-3349324

closes odoo/enterprise#51797

X-original-commit: f6f5342fc189e459481df0dae324099806c8496a
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>