[FIX] sale_subscription: fix the visibility for renewal quote stat button
As of now, the renewal quote stat button in the subscription module's sale order
form view is only visible when the sale order is in the draft state.

steps to reproduce:
- renew a so from the subscription module
- click on the renewal quote stat button and change the state=sent quotation
- the renewal button will disappear from the confirmed sale order form view

With this commit, the renewal quote stat button will now be displayed even when
the sale order is in the sent state.

task-3593973

closes odoo/enterprise#52992

X-original-commit: 3f8135bcb210d4b7290d0c324b349e31d1edc208
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>