[FIX] purchase_mrp_workoder_quality: operation from bom
Reminder of the behavior of boms :

When creating a manufacturing order using the bom of a product,
if the bom is modified after the creation of the manufacturing order,
it should not modify the manufacturing order.

This fix basically changes order of creation of objects in the demo data.
The work order has to be added before creating the manufacturing order
in order for it to appear on the manufacturing order.

odoo: https://github.com/odoo/odoo/pull/143496
opw-3568943

closes odoo/enterprise#52892

X-original-commit: b8c3e48b8771c6b91d89ae5a04f542d576ff9e0f
Related: odoo/odoo#146508
Signed-off-by: William Henrotin (whe) <whe@odoo.com>