[IMP] stock_barcode: don't count selected line as scanned line
Instead of consider a line was scanned when selected, we now consider a
line was scanned when it was modified (the line was marked as dirty.)

The purpose of this change is to avoid to change the destination of a
not scanned line who has been clicked but isn't selected anymore, while
still be able to change the destination of a not scanned line if it is
the currently selected line.
Indeed, when a destination or a package is assigned to a group of lines,
it's assigned to all previously scanned lines. If we consider a clicked
line is also scanned, that would mean we will change the destination of
line the user clicked on accidentally for example, which is not wanted.

closes odoo/enterprise#54002

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>