[IMP] documents: adapt document view from mail activity
Purpose
=======

When you receive an e-mail about an action to be carried out on a
document, the link in the message points to the document's
form view.

Unfortunately, this view doesn't allow you to access some
functionalities that would enable you to perform the requested action:
the preview, the inspector, the pdf manager...

But the kaban view does.

Specifications
==============

When users click on the link they are redirected to the kanban view,
previewing the document to be processed.

task-3476017

closes odoo/enterprise#51874

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>