[IMP] social_linkedin: migrate to the newest API version
Purpose
=======
On the 3 December 2023, we are using the API version 202211. However,
the version 202209 is depreciated, so it might be the time to upgrade.

Upgrade to the version 202401.

Technical
=========
The only technical change is the fact that projections are removed.

Those were useful to fetch everything we needed in a single HTTP request,
and so now, we have to make many HTTP requests to fetch the data (at
least, one per "model").

Here are the use case impacted:
- Fetching the pages required one call before, now it needs 3
  (one to fetch the pages for which the user is admin, one to fetch
  some information about the pages and one to fetch the images)
- Fetching comments needed  1 HTTP call (and everything was fetched
  using the projection feature), now it needs 4 (one call to get the
  list of comments, one call to get the pages information, one call
  to get the users information, one call to get the profile picture
  of the users / pages)

Task-3620605

X-original-commit: ab5f5e27f2e3cac376fb9eb7252b18d7d54ea24e
Part-of: odoo/enterprise#56316