[FIX] account_accountant: fix reconciliation write-off on same account
Steps to reproduce:
- Install Accounting
- Create a customer payment of $100
- Go to "Accounting / Accounting / Journals / Journal Items"
- Set a Group By Account
- Under "Outstanding Receipts" account, you should find the payment
- Select the payment and reconcile it
- Use the same account for the write-off (i.e. Outstanding Receipts)
- Reconcile
=> The payment and the write-off are partially reconciled.
They should be fully reconciled.

Cause:
The write-off is created with "Outstanding Receipts" account for both the write-off
line and the counterpart line.
Once created, a filter on the reconciliation account (i.e. Outstanding Receipts)
is performed on the write-off lines to reconciled them with the payment line.
As both write-off lines are using the same account, we try to reconcile the
payment line ($100), the write-off line ($100) and its counterpart ($-100) together,
which cannot be fully reconciled.

Solution:
The first line of the write-off should contain the counterpart line.
So we can just select that line for reconciliation.

opw-3553008

closes odoo/enterprise#53976

X-original-commit: ca30f8d848979ef7305a093447269a2112e4f498
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>