[FIX] whatsapp: skip UserError in batch whatsapp composer usage
Steps to reproduce:
1. Install the WhatsApp, contact module.
2. Create a WhatsApp template for model contact with phone field = mobile.
3. Click on Allow Multi and refresh the page.
4. Navigate to the list view of Contacts and select multiple records that
   don't have mobile available.
5. Go to actions > whatsapp message
This results in a UserError instead of opening a composer

Solution:
Don't validate the number in case of multi for WhatsApp composer. Also fix
'invalid_number' count: a wrong number should be considered as invalid
and not crash.

Task-3527702
Sub-part of Task-3608129 (Whatsapp: Moultifix !)

X-original-commit: 4c2f2f389c829d69a66440f8acc21ab1f0e40dcd
Part-of: odoo/enterprise#52905