[FIX] industry_fsm_report: fix access right error when trying to open url of the task
Before this commit, when the internal user tries to access to the form view of a task
with worksheet template set, he got an access right error because he cannot access to
the worksheet model linked to the task. The reason is because worksheet_count is computed
by counting the number of worksheets linked to the task when the worksheet template is
set on that task and the internal user without any access to Field Service and Project
apps cannot access to worksheet models.

This commit allows to compute `worksheet_count` in sudo since the data is not really
confidential and so no need to apply the access rights to get the number of worksheets
linked to a task. By doing that, the internal user without any access to Field Service
and Project apps can access to the form view of task without any issue.

task-3430762

closes odoo/enterprise#54036

X-original-commit: d983418e4b7266e9e252aa83aae144101d34c2a3
Related: odoo/odoo#148837
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>