[FIX] documents: close the preview while restoring the documents
How to reproduce:
1. Select a file/doc.
2. Move to trash ( by trash bin icon).
3. Go to trash workspace.
4. Preview the file/doc.
5. Restore the file.
6. Now, try to preview it from the inspector.
7. It throws an error

Technical Reason:
Before version 17.0, Inside "_toggleArchive", we were triggering the event to close the preview,
but In 17.0 "_toggleArchive" is removed after this commit , So the preview remains open.
https://github.com/odoo-dev/enterprise/commit/2254eb07fd42ec88341d9d7ac92f09a379c6fa55#diff-78d9f8faf609f35072d4b2617a9e13b3f137a679efb2c46c84acf29da4edfe46L426

After this commit:
No traceback will appear, and the preview will be closed,
after clicking on the restore button.

Task-3619214

closes odoo/enterprise#52226

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>