[FIX] planning: allow flexible hours
Issue Description:
==================
An employee is configured for flexible work hours. However, when their schedule is planned, it fails to display correctly.

In Odoo 16.3, to establish flexible hours, one must select `planning` in the employee`s contract. Consequently, the `Working hours` field (found under `Schedule` in the `Work Information` tab) should be left blank. Intriguingly, if `Working hours` are reinstated in `Work Information,` the planning becomes visible, but this action contradicts the flexible hours setting.

Steps to Reproduce:
===================
1. Navigate to the `Employees` app and select an employee.
2. Access the `Work Information` tab and clear the `Working hours` field.
3. Go to the `Planning` app, and publish a schedule for the selected employee.
4. Switch to developer mode and proceed to `Settings` > `Technical` > `Email` > `Emails`.
5. Attempt to open a link from the relevant email; the planning will not be displayed.

Proposed Solution:
==================
Modify the process so that when an employee`s `Working Hours` are not set, the system bypasses the calculation of 'minTime' and 'maxTime' values in the 'timeGridDay' and 'timeGridWeek' components. This should allow for the correct display of planning schedules for employees with flexible working hours.

opw-3608860

closes odoo/enterprise#52123

X-original-commit: f26eb9b7d38434e48179c0b2c038cfdcb06a36d8
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Ilya Rudy (ilru) <ilru@odoo.com>