[FIX] whatsapp: prevent traceback when removing {{1}} from template body
When `{{1}}` is removed from the template body and there are multiple variables
a traceback appears.

Steps to reproduce:
1. Create a template with the body `Hello, {{2}} Thanks! {{3}}`.
2. Traceback appears

Problem:
`_check_body_variables` updates the `missing` variable if the length of
variables is greater than one. the `next` function to update the `missing`
variable by checking the proper sequence of `variable_indices`.  In this case
`variable_indices` will be `[2, 3]` which is a proper sequence. Due to this
`next` will get nothing and it will throw a `StopIteration` error.

Solution:
Statically, `missing` is assigned to 1 so there is no need to update `missing`
if `variable_indices` doesn't start from one.

Task-3565540
Sub-part of Task-3608129 (Whatsapp: Moultifix !)

X-original-commit: c477198b1abdff9e82e5f546c26911340c415fbc
Part-of: odoo/enterprise#52840