[FIX] whatsapp: fix non 'free_text' text header variable
This error is produced when a template is sent whose `header_text` contains a
variable and no template variables are set to `free text`.

Steps to reproduce:
1. Create a template with `header text` set as `Hello {{1}}`
2. Set all variable types to any other than `free text`
3. Try to send this template using the composer
4. Traceback is produced

Problem:
The function `_get_header_component` sets the value of header parameters using
`free_text_json.get('header_text')`. As there is no `free text` in the template
`free_text_json` will be `False`, which will throw an error of not having a`get`
attribute.

Solution:
Be defensive when accessing 'free_text_json'.

Task-3602854
Sub-part of Task-3608129 (Whatsapp: Moultifix !)

X-original-commit: 9fd30abcb56b537007f907746a2ff023e0fc4c4c
Part-of: odoo/enterprise#52840