[IMP] spreadsheet_edition: restore version
This commit adds the ability to restore a spreadsheet to a previous version.
It's implemented by deleting all revisions that occured after the revision
the user wants to restore.

We could also preserves the full history and adds a revision like "restore"
on top of the history, but that would be technically challenging given how
the collaborative history works.

Notes on concurrency
--------------------

We must avoid corrupted history if a user restores a spreadsheet to a previous
version while another user concurrently update that spreadsheet.

Let's say we start with the following history: A B C D E

Alice restores the spreadsheet to revision B
=> the history should become A B

Let's say Bob concurrently updates the spreadsheet with revision F, based on
revision E.
=> the history should become A B C D E F

We must avoid the final history to look like A B F!
Fortunatly, it cannot happen thanks to the unique partial indexes (see 2
commits before this one)

closes odoo/enterprise#51489

Task: 3611266
Related: odoo/upgrade#5410
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>