[FIX] whatsapp: post rejection only if template is rejected
The "Your Template has been rejected." message is posted every time even if it
is not rejected when template status is updated by webhook

Reason:
In the webhook post controller when there is any description in
message_template_status_update webhook event then it posts the rejection
message without checking the value of the event.

Solution:
Post rejection message only if the value of the event is `rejected`.

Task-3516056
Sub-part of Task-3608129 (Whatsapp: Moultifix !)

X-original-commit: 2770ced1ae1b2074145a887820a3abccf2c6ee5b
Part-of: odoo/enterprise#52905