[FIX] knowledge: prevent zeroWidthSpace in /file name
There is an issue on some browsers where the name of a download file always has
an `_` at the end of it.

How to reproduce:
- open Knowledge and insert a /file block
- click on the file name and edit it (don't use the `_` character)
- confirm with enter
- click on Download

Current Behavior:
- the downloaded file name has an `_` at the end of it even though it should not

Technical explanation:
The pen icon that is displayed on hover contains a `zeroWidthSpace` character
added by the editor and that character was mistakenly appended to the name of
the file since it was recovered with `textContent` applied on a `div` ancestor.

The name will now be recovered directly from the `textContent` of the `span`,
and a class is added to the `span` to ensure that it will not be sanitized by
the editor (see [1]).

[1]: https://github.com/odoo/odoo/commit/0f4914a1576108bb0843694a62d679bc78a8574a

task-3627975

closes odoo/enterprise#52738

X-original-commit: 20a1279fa629022b836327338b38415a00f5a0bc
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>