[FIX] whatsapp_*: correctly set category in manifest
A 'WhatsApp' module category exists in data. Let us use it in whatsapp
modules manifest.

Note that since odoo/odoo#55907 categories without modules belonging to
child categories are not displayed in the search panel, not sure why (specs
are not really clear about that point). Anyway not the scope of this task
to fix that, tant pis hein.

Courtesy of Niyas Raphy (sorry, lost my get-br script for authorship)

Task-3498748
Sub-part of Task-3608129 (Whatsapp: Moultifix !)

X-original-commit: 2a5950cb750c055662ce84e548da5ce88c56b5d4
Part-of: odoo/enterprise#52840