[FIX] whatsapp: improve 'phone' computation and context support on composer
Steps to reproduce:
1. Install `Contacts` and open any contact
2. Hover over the phone or mobile field and click on WhatsApp
3. Remove the selected template
4. Phone number will be removed

Problem:
When the composer is opened using the phone or mobile field, the composer
phone field should be automatically set based on the `default_phone` value from
the context. In the `_compute_number` function, the `phone` variable is
initially set to None and is updated only if the `default_phone` context is
not present.

Solution:
Better compute 'phone', based on batch_mode, then on available context value.

This commit also solves an error when the template is removed from the composer
which is opened using chatter.

This commit also solves the "... (X Others)" display that was incorrectly
computed as always equal to the number of found numbers, instead of number
of left numbers.

Task-3535128
Sub-part of Task-3608129 (Whatsapp: Moultifix !)

X-original-commit: ad48d2a23c87f28e87eca90900d6ad6c266afa43
Part-of: odoo/enterprise#52905