[FIX] l10n_rs_reports: Fix unbalanced Balance Sheet
According to the Balance Sheet diagnosis tool in #36838, the Balance
Sheet was unbalanced. On closer inspection, this was seen to be caused
by an incorrect cross-report to the Profit and Loss in the Current Year
Earnings and Retained Earnings sections.

The same issue is present all the way to master.

This commit fixes the cross-report references. We also take the
opportunity to remove all the redundant `date_scope=normal` data.
Since Balance Sheet reports work with a single (end) date, the normal
date_scope works exactly the same as the strict_range date_scope, which
is the default.

closes odoo/enterprise#52184

Taskid: 3060790
X-original-commit: 517f7cd62f23ab3180f2b443a250fe3ed3657383
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>