[IMP] hr_referral: start using ribbons in applicants kanban
Currently we show the status of the applicant as a simple tag on the top right
of the kanban view. However, at Odoo we normally use ribbons for these cases.
This commit improves the kanban view of the applicant by switching from tags to
ribbons. Also, we rename the hr_applicant referral state from "In progress" to
"Ongoing" as the former was taking too much space and overflowing out of the
ribbon.

task-3621480

closes odoo/enterprise#51972

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>